Overhaul of serial port objects
Before, there was only one serial port, the serial port currently
connected to. Now there is an array 'sPorts', each entry being a serial
port. This allows the program to open multiple serial ports
simultaneously and scan them in parrallel. This is needed because the
time from opening the port until response on the arm now varies between
100ms (Non bootloader/no reset jumper)and ~6500ms (bootloader with
reset jumper)