Emergency Stop, pose clearing and export fixes
Added a button for the emergency stop function (extended instruction 17)
Poses now clear when orientation or modes are changed to avoid wrong
data  from getting piped to the wrong mode
File export now has a dialog to chose a folder for poses.h. Also, all
variables are not correctly offer.